requirements.txt
python-binance>=1.0.19
aiohttp>=3.9.0
fastapi>=0.104.0
uvicorn>=0.24.0
loguru>=0.7.0
python-dotenv>=1.0.0
